==> Notebooks/00_Data_Extraction.ipynb <==
# 00.Data Extraction

!pip install openpyxl

import pandas as pd
import numpy as np
df = pd.read_excel('../Data/Precios_Coomproriente_05_06_25.xlsx')
Productos=df['Producto'].unique()
df.head(10)


print(f"Columnas del DataFrame:{Productos}")

## Extraer datos agronet

import pandas as pd
import glob
import os

# Ruta a la carpeta que contiene los archivos .xlsx
ruta = "../Data/Agronet"  

# Obtener la lista de todos los archivos .xlsx en la carpeta
archivos = glob.glob(os.path.join(ruta, "*.xlsx"))

# Verifica que haya archivos encontrados
if not archivos:
    print("No se encontraron archivos .xlsx en la ruta especificada.")
else:
    # Leer y concatenar todos los archivos en un solo DataFrame
    df_concatenado = pd.concat([pd.read_excel(archivo) for archivo in archivos], ignore_index=True)

    # Mostrar las primeras filas del DataFrame concatenado
    print("Primeras filas del DataFrame concatenado:")
    print(df_concatenado.head())

    # (Opcional) Guardar el DataFrame concatenado en un nuevo archivo Excel
    # df_concatenado.to_excel("df_concatenado.xlsx", index=False)
    df_concatenado.info()


df_concatenado['Mercado'].unique()

df_concatenado['Producto'].unique()


df_concatenado.to_excel("../Data/datos_agronet.xlsx", index=False)
